+++ weevent-broker/src/main/java/com/webank/weevent/broker/fisco/FiscoBcosTopicAdmin.java
@@ -8,6 +8,11 @@ import com.webank.weevent.broker.fisco.dto.ListPage;
 import com.webank.weevent.broker.fisco.util.ParamCheckUtils;
 import com.webank.weevent.broker.fisco.web3sdk.FiscoBcosDelegate;
 import com.webank.weevent.broker.plugin.IEventTopic;
+import com.webank.weevent.protocol.rest.entity.GroupGeneral;
+import com.webank.weevent.protocol.rest.entity.QueryEntity;
+import com.webank.weevent.protocol.rest.entity.TbBlock;
+import com.webank.weevent.protocol.rest.entity.TbNode;
+import com.webank.weevent.protocol.rest.entity.TbTransHash;
 import com.webank.weevent.sdk.BrokerException;
 import com.webank.weevent.sdk.ErrorCode;
 import com.webank.weevent.sdk.TopicInfo;
@@ -143,4 +148,24 @@ public class FiscoBcosTopicAdmin implements IEventTopic {
     public void validateGroupId(String groupId) throws BrokerException {
         ParamCheckUtils.validateGroupId(groupId, fiscoBcosDelegate.listGroupId());
     }
+    @Override
+    public GroupGeneral getGroupGeneral(String groupId) throws BrokerException {
+        return fiscoBcosDelegate.getGroupGeneral(groupId);
+    }
+
+    @Override
+    public List<TbTransHash> queryTransList(QueryEntity queryEntity) throws BrokerException {
+        return fiscoBcosDelegate.queryTransList(queryEntity);
+    }
+
+    @Override
+    public List<TbBlock> queryBlockList(QueryEntity queryEntity) throws BrokerException {
+        return fiscoBcosDelegate.queryBlockList(queryEntity);
+    }
+
+    @Override
+    public List<TbNode> queryNodeList(QueryEntity queryEntity) throws BrokerException {
+        return fiscoBcosDelegate.queryNodeList(queryEntity);
+    }
+
 }
