+++ weevent-broker/src/main/java/com/webank/weevent/protocol/rest/AdminRest.java
@@ -9,7 +9,6 @@ import java.util.Map;
 import com.webank.weevent.BrokerApplication;
 import com.webank.weevent.broker.config.BuildInfo;
 import com.webank.weevent.broker.fisco.util.SystemInfoUtils;
-import com.webank.weevent.broker.plugin.IAdmin;
 import com.webank.weevent.broker.plugin.IConsumer;
 import com.webank.weevent.protocol.rest.entity.GroupGeneral;
 import com.webank.weevent.protocol.rest.entity.QueryEntity;
@@ -27,7 +26,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.http.client.SimpleClientHttpRequestFactory;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
@@ -43,7 +41,6 @@ import org.springframework.web.client.RestTemplate;
 @Slf4j
 public class AdminRest extends RestHA {
     private IConsumer consumer;
-    private IAdmin admin;
     private BuildInfo buildInfo;
 
     @Autowired
@@ -54,11 +51,6 @@ public class AdminRest extends RestHA {
     @Autowired
     public void setBuildInfo(BuildInfo buildInfo) {
         this.buildInfo = buildInfo;
-    }
-
-    @Autowired
-    public void setAdmin(IAdmin admin) {
-        this.admin = admin;
     }
 
     @RequestMapping(path = "/listSubscription")
@@ -112,14 +104,14 @@ public class AdminRest extends RestHA {
     /**
      * get general
      */
-    @RequestMapping(path = "/group/general/{groupId}")
-    public ResponseData getGroupGeneral(@PathVariable("groupId") Integer groupId) throws BrokerException {
+    @RequestMapping(path = "/group/general")
+    public ResponseData getGroupGeneral(@RequestParam("groupId") Integer groupId) throws BrokerException {
         ResponseData responseData = new ResponseData();
         Instant startTime = Instant.now();
         log.info("start getGroupGeneral startTime:{} groupId:{}", startTime.toEpochMilli(),
                 groupId);
 
-        GroupGeneral groupGeneral = this.admin.getGroupGeneral(String.valueOf(groupId));
+        GroupGeneral groupGeneral = this.consumer.getGroupGeneral(String.valueOf(groupId));
         responseData.setCode(ErrorCode.SUCCESS.getCode());
         responseData.setMessage(ErrorCode.SUCCESS.getCodeDesc());
         responseData.setData(groupGeneral);
@@ -130,10 +122,10 @@ public class AdminRest extends RestHA {
     /**
      * query transaction list.
      */
-    @RequestMapping(path = "/transaction/transList/{groupId}/{pageNumber}/{pageSize}")
-    public ResponseData queryTransList(@PathVariable("groupId") Integer groupId,
-                                       @PathVariable("pageNumber") Integer pageNumber,
-                                       @PathVariable("pageSize") Integer pageSize,
+    @RequestMapping(path = "/transaction/transList")
+    public ResponseData queryTransList(@RequestParam("groupId") Integer groupId,
+                                       @RequestParam("pageNumber") Integer pageNumber,
+                                       @RequestParam("pageSize") Integer pageSize,
                                        @RequestParam(value = "transactionHash", required = false) String transHash,
                                        @RequestParam(value = "blockNumber", required = false) BigInteger blockNumber)
             throws BrokerException {
@@ -147,7 +139,7 @@ public class AdminRest extends RestHA {
         ResponseData responseData = new ResponseData();
         QueryEntity queryEntity = new QueryEntity(groupId.toString(), pageNumber, pageSize, transHash, blockNumber);
 
-        List<TbTransHash> tbTransHashes = this.admin.queryTransList(queryEntity);
+        List<TbTransHash> tbTransHashes = this.consumer.queryTransList(queryEntity);
         responseData.setCode(ErrorCode.SUCCESS.getCode());
         responseData.setMessage(ErrorCode.SUCCESS.getCodeDesc());
         responseData.setData(tbTransHashes);
@@ -159,10 +151,10 @@ public class AdminRest extends RestHA {
     /**
      * query block list.
      */
-    @RequestMapping(path = "/block/blockList/{groupId}/{pageNumber}/{pageSize}")
-    public ResponseData queryBlockList(@PathVariable("groupId") Integer groupId,
-                                       @PathVariable("pageNumber") Integer pageNumber,
-                                       @PathVariable("pageSize") Integer pageSize,
+    @RequestMapping(path = "/block/blockList")
+    public ResponseData queryBlockList(@RequestParam("groupId") Integer groupId,
+                                       @RequestParam("pageNumber") Integer pageNumber,
+                                       @RequestParam("pageSize") Integer pageSize,
                                        @RequestParam(value = "pkHash", required = false) String pkHash,
                                        @RequestParam(value = "blockNumber", required = false) BigInteger blockNumber)
             throws BrokerException {
@@ -176,7 +168,7 @@ public class AdminRest extends RestHA {
         ResponseData responseData = new ResponseData();
         QueryEntity queryEntity = new QueryEntity(groupId.toString(), pageNumber, pageSize, pkHash, blockNumber);
 
-        List<TbBlock> tbBlocks = this.admin.queryBlockList(queryEntity);
+        List<TbBlock> tbBlocks = this.consumer.queryBlockList(queryEntity);
         responseData.setCode(ErrorCode.SUCCESS.getCode());
         responseData.setMessage(ErrorCode.SUCCESS.getCodeDesc());
         responseData.setData(tbBlocks.toArray());
@@ -187,10 +179,10 @@ public class AdminRest extends RestHA {
     /**
      * qurey node info list.
      */
-    @RequestMapping(path = "/node/nodeList/{groupId}/{pageNumber}/{pageSize}")
-    public ResponseData queryNodeList(@PathVariable("groupId") Integer groupId,
-                                      @PathVariable("pageNumber") Integer pageNumber,
-                                      @PathVariable("pageSize") Integer pageSize,
+    @RequestMapping(path = "/node/nodeList")
+    public ResponseData queryNodeList(@RequestParam("groupId") Integer groupId,
+                                      @RequestParam("pageNumber") Integer pageNumber,
+                                      @RequestParam("pageSize") Integer pageSize,
                                       @RequestParam(value = "nodeName", required = false) String nodeName)
             throws BrokerException {
         Instant startTime = Instant.now();
@@ -203,7 +195,7 @@ public class AdminRest extends RestHA {
         ResponseData responseData = new ResponseData();
         responseData.setCode(ErrorCode.SUCCESS.getCode());
         responseData.setMessage(ErrorCode.SUCCESS.getCodeDesc());
-        List<TbNode> tbNodeList = this.admin.queryNodeList(queryEntity);
+        List<TbNode> tbNodeList = this.consumer.queryNodeList(queryEntity);
         responseData.setTotalCount(tbNodeList == null ? 0 : tbNodeList.size());
         responseData.setData(tbNodeList);
         return responseData;
@@ -212,8 +204,8 @@ public class AdminRest extends RestHA {
     /**
      * query the number of transactions in the last week
      */
-    @GetMapping("/group/transDaily/{groupId}")
-    public ResponseData getTransDaily(@PathVariable("groupId") Integer groupId)
+    @GetMapping("/group/transDaily")
+    public ResponseData getTransDaily(@RequestParam("groupId") Integer groupId)
             throws BrokerException {
         ResponseData responseData = new ResponseData();
         responseData.setCode(ErrorCode.SUCCESS.getCode());
